Move asset endpoint to the FileResource
